feat: less custom UI
instead of replacing the whole work edit widget, replace only the edit button to open our custom dialog.

Fixes #88, #92
