+++ scripts/acum-work-import/src/import-album.ts
@@ -2,13 +2,21 @@ import {Creator, Creators, Entity, entityUrl, fetchWorks, IPBaseNumber, trackNam
 import {linkArtists} from '#artists.ts';
 import {addArrangerRelationship, addWriterRelationship} from '#relationships.ts';
 import {AddWarning} from '#ui/warnings.tsx';
-import {workEditDataEqual} from '#work-edit-data.ts';
+import {addWorkEditor} from '#ui/work-editor.tsx';
+import {workEditData, workEditDataEqual} from '#work-edit-data.ts';
 import {WorkStateWithEditDataT} from '#work-state.ts';
-import {addWork, linkWriters} from '#works.ts';
+import {createNewWork, linkWriters} from '#works.ts';
 import {head} from '@repo/common/head';
-import {compareInsensitive} from '@repo/musicbrainz-ext/compare';
+import {
+  compareInsensitive,
+  compareNumbers,
+  compareTargetTypeWithGroup,
+  compareWorks,
+} from '@repo/musicbrainz-ext/compare';
+import {REL_STATUS_REMOVE} from '@repo/musicbrainz-ext/constants';
 import {addEditNote} from '@repo/musicbrainz-ext/edit-note';
 import {trackRecordingState} from '@repo/musicbrainz-ext/track-recording-state';
+import {iterateRelationshipsInTargetTypeGroup} from '@repo/musicbrainz-ext/type-group';
 import {
   connect,
   count,
@@ -36,6 +44,8 @@ import {
   MediumRecordingStateT,
   MediumRecordingStateTreeT,
   MediumWithRecordingsT,
+  MediumWorkStateT,
+  MediumWorkStateTreeT,
   RecordingT,
   ReleaseRelationshipEditorStateT,
 } from 'typedbrainz/types';
@@ -94,7 +104,6 @@ async function importSelectedWorks(
               selectedRecording.workBean,
               selectedRecording.recordingState.recording,
               await addWork(
-                selectedRecording.position,
                 selectedRecording.index,
                 selectedRecording.workBean,
                 selectedRecording.recordingState,
@@ -253,4 +262,76 @@ function selectedMediums(entity: Entity, noSelection: boolean, addWarning: AddWa
   }
 
   return selected;
+}
+
+async function addWork(
+  index: number | undefined,
+  track: WorkBean,
+  recordingState: MediumRecordingStateT,
+  addWarning: AddWarning
+): Promise<WorkStateWithEditDataT> {
+  const workState = (await (async () => {
+    if (!MB || !MB.tree || !isReleaseRelationshipEditor(MB?.relationshipEditor)) {
+      throw new Error('MB or MB.tree is not defined or not a release relationship editor');
+    }
+
+    const existing = relatedWork(recordingState.relatedWorks);
+    if (existing) {
+      return existing as WorkStateWithEditDataT;
+    }
+
+    const newWork = await createNewWork(index, track, recordingState);
+
+    recordingState = refreshRecordingState(recordingState.recording);
+
+    return MB.tree.find(
+      recordingState.relatedWorks,
+      newWork,
+      (work, relatedWork) => compareWorks(work, relatedWork.work),
+      null
+    )!;
+  })()) as WorkStateWithEditDataT;
+
+  Object.assign(workState, await workEditData(workState.work, track, addWarning));
+  addWorkEditor(workState, recordingState);
+  return workState;
+}
+
+function refreshRecordingState(recording: RecordingT): MediumRecordingStateT {
+  if (!MB || !MB.tree || !isReleaseRelationshipEditor(MB?.relationshipEditor)) {
+    throw new Error('MB or MB.tree is not defined or not a release relationship editor');
+  }
+
+  const mediumRecordingStates = MB.tree.find(
+    MB.relationshipEditor.state.mediums,
+    MB.relationshipEditor.state.mediumsByRecordingId.get(recording.id)![0],
+    (mediumKey, [mediumVal]) => {
+      return compareNumbers(mediumKey?.id ?? 0, mediumVal.id);
+    },
+    null
+  )![1];
+  return MB.tree.find(
+    mediumRecordingStates,
+    recording,
+    (recording, recordingState) => compareNumbers(recording.id, recordingState.recording.id),
+    null
+  )!;
+}
+
+function relatedWork(relatedWorks: MediumWorkStateTreeT): MediumWorkStateT | undefined {
+  if (!MB?.tree) {
+    return;
+  }
+
+  const relatedWork = head(MB.tree.iterate(relatedWorks));
+  if (relatedWork) {
+    const targetTypeGroup = MB.tree.find(relatedWork.targetTypeGroups, 'recording', compareTargetTypeWithGroup, null);
+    if (targetTypeGroup) {
+      for (const relationship of iterateRelationshipsInTargetTypeGroup(targetTypeGroup)) {
+        if (relationship._status !== REL_STATUS_REMOVE) {
+          return relatedWork;
+        }
+      }
+    }
+  }
 }
