+++ scripts/acum-work-import/src/ui/work-editor.tsx
@@ -1,11 +1,9 @@
 import {useWorkEditData, WorkEditDataProvider} from '#ui/work-edit-data-provider.tsx';
 import {WorkEditDialog} from '#ui/work-edit-dialog.tsx';
 import {WorkStateWithEditDataT} from '#work-state.ts';
-import {Button} from '@kobalte/core/button';
-import {createEffect, createSignal, JSX, Show} from 'solid-js';
+import {createEffect, createSignal, Show} from 'solid-js';
 import {render} from 'solid-js/web';
-import {isReleaseRelationshipEditor} from 'typedbrainz';
-import {MediumRecordingStateT, WorkT} from 'typedbrainz/types';
+import {MediumRecordingStateT, RecordingT, WorkT} from 'typedbrainz/types';
 
 function isNewWork(work: WorkT) {
   return !work.gid;
@@ -15,40 +13,26 @@ function workLink(work: WorkT) {
   return isNewWork(work) ? `#new-work-${work.id}` : `/work/${work.gid}`;
 }
 
+function recordingLink(recording: RecordingT) {
+  return '/recording/' + recording.gid;
+}
+
 function WorkEditor(props: {
   workState: WorkStateWithEditDataT;
   recordingState: MediumRecordingStateT;
   originalHeader: HTMLHeadingElement;
 }) {
-  const isNew = isNewWork(props.workState.work);
-  const {isModified, workName} = useWorkEditData();
+  const {isModified} = useWorkEditData();
   const [isPending, setIsPending] = createSignal(isModified());
 
-  const selectRecording: JSX.ChangeEventHandler<HTMLInputElement, Event> = event => {
-    if (MB && isReleaseRelationshipEditor(MB.relationshipEditor)) {
-      MB.relationshipEditor.dispatch({
-        isSelected: event.currentTarget.checked,
-        work: props.workState.work,
-        type: 'toggle-select-work',
-      });
-    }
-  };
+  createEffect(() => {
+    const workLinkElement = props.originalHeader.querySelector<HTMLAnchorElement>('a[href^="/work/"]');
+    workLinkElement?.classList.toggle('rel-edit', isModified() && isPending());
+  });
 
-  const removeWork: JSX.EventHandler<HTMLButtonElement, MouseEvent> = () => {
-    setIsPending(false);
-    if (MB && isReleaseRelationshipEditor(MB.relationshipEditor)) {
-      MB.relationshipEditor.dispatch({
-        recording: props.recordingState.recording,
-        type: 'remove-work',
-        workState: props.workState,
-      });
-    }
-  };
-
-  createEffect((previousDisplay?: string) => {
-    const originalDisplay = props.originalHeader.style.display;
-    props.originalHeader.style.display = isPending() ? 'none' : (previousDisplay ?? '');
-    return originalDisplay;
+  createEffect(() => {
+    const editButton = props.originalHeader.querySelector<HTMLButtonElement>(':scope > button.edit-item');
+    editButton?.classList.toggle('pending', isPending());
   });
 
   document.getElementById('acum-work-cancel')?.addEventListener('click', () => {
@@ -57,28 +41,20 @@ function WorkEditor(props: {
 
   return (
     <Show when={isPending()}>
-      <input type="checkbox" class="work" checked={props.workState.isSelected} onChange={selectRecording} />{' '}
-      <Button class="icon remove-item" onClick={removeWork} />
       <WorkEditDialog onSubmit={() => setIsPending(false)} />{' '}
-      <a
-        href={workLink(props.workState.work)}
-        classList={{
-          'wrap-anywhere': true,
-          'rel-add': isNew,
-          'rel-edit': !isNew,
-        }}
-      >
-        {workName()}
-      </a>
     </Show>
   );
 }
 
 export function addWorkEditor(workState: WorkStateWithEditDataT, recordingState: MediumRecordingStateT) {
-  const header = document.querySelector(`.works a[href="${workLink(workState.work)}"]`)?.closest('h3');
-  if (header && !header.classList.contains('edit-work-button-container')) {
-    const container = (<h3 class="edit-work-button-container"></h3>) as HTMLHeadingElement;
-    header.parentElement?.insertBefore(container, header);
+  const track = document.querySelector(`.track:has(a[href="${recordingLink(recordingState.recording)}"])`);
+  const header = track?.querySelector<HTMLHeadingElement>(
+    `.works h3:has(a[href="${workLink(workState.work)}"]):not(:has(div.edit-work-button-container))`
+  );
+  if (header) {
+    const container = (<div class="edit-work-button-container"></div>) as HTMLDivElement;
+    const removeButton = header.querySelector<HTMLButtonElement>('button.remove-item');
+    removeButton?.insertAdjacentElement('afterend', container);
     render(
       () => (
         <WorkEditDataProvider workState={workState}>
