+++ scripts/acum-work-import/src/works.ts
@@ -3,57 +3,23 @@ import {trackName, WorkBean, workISWCs, workType} from '#acum.ts';
 import {linkArtists} from '#artists.ts';
 import {createRelationshipState} from '#relationships.ts';
 import {shouldSearchWorks} from '#ui/settings.tsx';
-import {AddWarning} from '#ui/warnings.tsx';
-import {addWorkEditor} from '#ui/work-editor.tsx';
-import {workEditData} from '#work-edit-data.ts';
-import {WorkStateWithEditDataT} from '#work-state.ts';
-import {head} from '@repo/common/head';
-import {compareNumbers, compareTargetTypeWithGroup, compareWorks} from '@repo/musicbrainz-ext/compare';
 import {
   COMPOSER_LINK_TYPE_ID,
   LYRICIST_LINK_TYPE_ID,
   MEDLEY_LINK_TYPE_ID,
   RECORDING_OF_LINK_TYPE_ID,
   REL_STATUS_ADD,
-  REL_STATUS_REMOVE,
   TRANSLATOR_LINK_TYPE_ID,
   WRITER_LINK_TYPE_ID,
 } from '@repo/musicbrainz-ext/constants';
 import {fetchJSON, tryFetchJSON} from '@repo/musicbrainz-ext/fetch';
 import {formatISWC} from '@repo/musicbrainz-ext/format-iswc';
 import {IswcLookupResultsT, WorkLookupResultT, WorkSearchResultsT} from '@repo/musicbrainz-ext/search-results';
-import {iterateRelationshipsInTargetTypeGroup} from '@repo/musicbrainz-ext/type-group';
 import {defaultIfEmpty, filter, firstValueFrom, from, mergeMap} from 'rxjs';
 import {isReleaseRelationshipEditor} from 'typedbrainz';
-import {
-  ArtistT,
-  LinkAttrT,
-  MediumRecordingStateT,
-  MediumWorkStateT,
-  MediumWorkStateTreeT,
-  RecordingT,
-  WorkT,
-} from 'typedbrainz/types';
+import {ArtistT, LinkAttrT, MediumRecordingStateT, WorkT} from 'typedbrainz/types';
 
 const workCache = new Map<string, WorkT>();
-
-function relatedWork(relatedWorks: MediumWorkStateTreeT): MediumWorkStateT | undefined {
-  if (!MB?.tree) {
-    return;
-  }
-
-  const relatedWork = head(MB.tree.iterate(relatedWorks));
-  if (relatedWork) {
-    const targetTypeGroup = MB.tree.find(relatedWork.targetTypeGroups, 'recording', compareTargetTypeWithGroup, null);
-    if (targetTypeGroup) {
-      for (const relationship of iterateRelationshipsInTargetTypeGroup(targetTypeGroup)) {
-        if (relationship._status !== REL_STATUS_REMOVE) {
-          return relatedWork;
-        }
-      }
-    }
-  }
-}
 
 export async function findWork(track: WorkBean) {
   const workId = await (async () => {
@@ -89,63 +55,7 @@ export async function findWork(track: WorkBean) {
   return undefined;
 }
 
-export async function addWork(
-  position: number,
-  index: number | undefined,
-  track: WorkBean,
-  recordingState: MediumRecordingStateT,
-  addWarning: AddWarning
-): Promise<WorkStateWithEditDataT> {
-  const workState = (await (async () => {
-    if (!MB || !MB.tree || !isReleaseRelationshipEditor(MB?.relationshipEditor)) {
-      throw new Error('MB or MB.tree is not defined or not a release relationship editor');
-    }
-
-    const existing = relatedWork(recordingState.relatedWorks);
-    if (existing) {
-      return existing as WorkStateWithEditDataT;
-    }
-
-    const newWork = await createNewWork(position, index, track, recordingState);
-
-    recordingState = refreshRecordingState(recordingState.recording);
-
-    return MB.tree.find(
-      recordingState.relatedWorks,
-      newWork,
-      (work, relatedWork) => compareWorks(work, relatedWork.work),
-      null
-    )!;
-  })()) as WorkStateWithEditDataT;
-
-  Object.assign(workState, await workEditData(workState.work, track, addWarning));
-  addWorkEditor(workState, recordingState);
-  return workState;
-}
-
-function refreshRecordingState(recording: RecordingT): MediumRecordingStateT {
-  if (!MB || !MB.tree || !isReleaseRelationshipEditor(MB?.relationshipEditor)) {
-    throw new Error('MB or MB.tree is not defined or not a release relationship editor');
-  }
-
-  const mediumRecordingStates = MB.tree.find(
-    MB.relationshipEditor.state.mediums,
-    MB.relationshipEditor.state.mediumsByRecordingId.get(recording.id)![0],
-    (mediumKey, [mediumVal]) => {
-      return compareNumbers(mediumKey?.id ?? 0, mediumVal.id);
-    },
-    null
-  )![1];
-  return MB.tree.find(
-    mediumRecordingStates,
-    recording,
-    (recording, recordingState) => compareNumbers(recording.id, recordingState.recording.id),
-    null
-  )!;
-}
-
-async function createNewWork(
-  position: number,
+export async function createNewWork(
   index: number | undefined,
   track: WorkBean,
   recordingState: MediumRecordingStateT
@@ -208,13 +118,14 @@ async function createNewWork(
   // wait for the work to be added
   await new Promise<void>(resolve => {
     new MutationObserver((_mutations, observer) => {
+      console.log('work added observer', newWork);
       if (
         document.querySelector(`.works a[href="${newWork.gid ? `/work/${newWork.gid}` : `#new-work-${newWork.id}`}"]`)
       ) {
         observer.disconnect();
         resolve();
       }
-    }).observe(document.querySelector('.release-relationship-editor')!, {
+    }).observe(document.body, {
       childList: true,
       subtree: true,
     });
